backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "beautifulsoup4==4.13.4",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "fastapi-cli==0.0.7",
    "fastapi[standard]==0.115.12",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.25.0",
    "google-api-python-client==2.171.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-generativeai==0.8.5",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.3",
    "grpcio==1.72.1",
    "grpcio-status==1.71.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "openai==1.84.0",
    "playwright==1.52.0",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pyee==13.0.0",
    "pygments==2.19.1",
    "pyparsing==3.2.3",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "rich==14.0.0",
    "rich-toolkit==0.14.7",
    "rsa==4.9.1",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "starlette==0.46.2",
    "tqdm==4.67.1",
    "typer==0.16.0",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "uritemplate==4.2.0",
    "urllib3==2.4.0",
    "uvicorn==0.34.3",
    "uvloop==0.21.0",
    "watchfiles==1.0.5",
    "websockets==15.0.1",
]
